MonthPicker and its tests have been moved to a seperate file.
