+++ myApp/app/screens/CalendarFetching.js
@@ -22,6 +22,8 @@ import { Ionicons } from "@expo/vector-icons";
 import LayoutWrapper from "../components/LayoutWrapper.js";
 import HeaderButtons from "../components/HeaderButtons.js";
 import { Picker } from '@react-native-picker/picker';
+import MonthPicker from '../components/MonthPicker';
+
 
 export default function CalendarFetching() {
   const navigation = useNavigation();
@@ -194,32 +196,11 @@ export default function CalendarFetching() {
 
                 {/* Months Ahead Input */}
                 <View style={{ marginTop: 10 }}>
-                  <Text style={styles.subtitle}>Show events for the upcoming:</Text>
-                  <View
-                    style={{
-                      borderWidth: 1,
-                      borderColor: '#ccc',
-                      borderRadius: 6,
-                      overflow: 'hidden', 
-                      marginBottom: 10
-                    }}
-                  >
-                    <Picker
-                      testID="month-picker"
-                      selectedValue={monthsAhead}
-                      onValueChange={(itemValue) => setMonthsAhead(itemValue)}
-                      style={{ 
-                        height: 50 }}
-                    >
-                      <Picker.Item label="1 Month" value="1" testID="1m"/>
-                      <Picker.Item label="2 Months" value="2" testID="2m"/>
-                      <Picker.Item label="3 Months" value="3" testID="3m"/>
-                      <Picker.Item label="4 Months" value="4" testID="4m"/>
-                      <Picker.Item label="5 Months" value="5" testID="5m"/>
-                      <Picker.Item label="6 Months" value="6" testID="6m"/>
-                      <Picker.Item label="All Events" value="all" testID="allm"/>
-                    </Picker>
-                  </View>
+                  <MonthPicker
+                    monthsAhead={monthsAhead}
+                    setMonthsAhead={setMonthsAhead}
+                    styles={styles}
+                  />
                 </View>
 
                 {/* Connect Button */}
